+++ src/main/java/roomescape/config/interceptor/AdminHandlerInterceptor.java
@@ -4,8 +4,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.domain.LoginMember;
-import roomescape.domain.Role;
+import roomescape.entity.LoginMember;
+import roomescape.entity.Role;
 import roomescape.exception.ForbiddenException;
 import roomescape.jwt.JwtExtractor;
 import roomescape.service.AuthService;
