+++ src/main/java/roomescape/config/interceptor/CheckRoleHandlerInterceptor.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.annotation.CheckRole;
-import roomescape.domain.LoginMember;
-import roomescape.domain.Role;
+import roomescape.entity.LoginMember;
+import roomescape.entity.Role;
 import roomescape.exception.ForbiddenException;
 import roomescape.jwt.JwtExtractor;
 import roomescape.service.AuthService;
