+++ src/main/java/roomescape/config/resolver/LoginMemberArgumentResolver.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.domain.LoginMember;
+import roomescape.entity.LoginMember;
 import roomescape.jwt.JwtExtractor;
 import roomescape.service.AuthService;
