+++ src/main/java/roomescape/controller/AuthController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.LoginMember;
-import roomescape.domain.Member;
+import roomescape.entity.LoginMember;
+import roomescape.entity.Member;
 import roomescape.dto.request.LoginRequest;
 import roomescape.service.AuthService;
 import roomescape.service.MemberService;
