+++ src/main/java/roomescape/controller/ReservationController.java
@@ -13,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.annotation.CheckRole;
-import roomescape.domain.LoginMember;
-import roomescape.domain.Reservation;
-import roomescape.domain.Role;
+import roomescape.entity.LoginMember;
+import roomescape.entity.Reservation;
+import roomescape.entity.Role;
 import roomescape.dto.request.CreateReservationRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
