+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.annotation.CheckRole;
-import roomescape.domain.ReservationSlot;
-import roomescape.domain.ReservationSlots;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Role;
+import roomescape.entity.ReservationSlot;
+import roomescape.entity.ReservationSlots;
+import roomescape.entity.ReservationTime;
+import roomescape.entity.Role;
 import roomescape.dto.request.AvailableTimeRequest;
 import roomescape.dto.request.CreateReservationTimeRequest;
 import roomescape.dto.response.ReservationTimeResponse;
