+++ src/main/java/roomescape/entity/Reservation.java
@@ -1,20 +1,35 @@
-package roomescape.domain;
+package roomescape.entity;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import roomescape.exception.InvalidReservationException;
 
+@Entity
 public class Reservation {
 
-    private final Long id;
-    private final Member member;
-    private final LocalDate date;
-    private final ReservationTime time;
-    private final Theme theme;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    private Member member;
+
+    private LocalDate date;
+
+    @ManyToOne
+    private ReservationTime time;
+
+    @ManyToOne
+    private Theme theme;
+
+    public Reservation() {
+    }
 
     public Reservation(Long id, Member member, LocalDate date, ReservationTime time, Theme theme) {
-        validate(member, date, time, theme);
         this.id = id;
         this.member = member;
         this.date = date;
@@ -26,17 +41,6 @@ public class Reservation {
         this(null, member, date, time, theme);
     }
 
-    private void validate(Member member, LocalDate date, ReservationTime time, Theme theme) {
-        if (member == null) {
-            throw new InvalidReservationException("예약자는 비어있을 수 없습니다");
-        }
-        if (date == null || time == null) {
-            throw new InvalidReservationException("시간은 비어있을 수 없습니다.");
-        }
-        if (theme == null) {
-            throw new InvalidReservationException("테마는 비어있을 수 없습니다.");
-        }
-    }
 
     public boolean isBefore(LocalDateTime compareDateTime) {
         LocalDateTime reservationDateTime = LocalDateTime.of(date, getStartAt());
