+++ src/main/java/roomescape/entity/ReservationTime.java
@@ -1,28 +1,29 @@
-package roomescape.domain;
+package roomescape.entity;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import java.time.LocalTime;
 import java.util.Objects;
-import roomescape.exception.InvalidReservationTimeException;
 
+@Entity
 public class ReservationTime {
 
-    private final Long id;
-    private final LocalTime time;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private LocalTime time;
+
+    public ReservationTime() {
+    }
 
     public ReservationTime(Long id, LocalTime time) {
-        validate(time);
         this.id = id;
         this.time = time;
     }
 
     public ReservationTime(LocalTime time) {
         this(null, time);
-    }
-
-    private void validate(LocalTime time) {
-        if (time == null) {
-            throw new InvalidReservationTimeException("유효하지 않은 예약시간입니다.");
-        }
     }
 
     public Long getId() {
