+++ src/main/java/roomescape/entity/Theme.java
@@ -1,17 +1,24 @@
-package roomescape.domain;
+package roomescape.entity;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import java.util.Objects;
-import roomescape.exception.InvalidThemeException;
 
+@Entity
 public class Theme {
 
-    private final Long id;
-    private final String name;
-    private final String description;
-    private final String thumbnail;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private String description;
+    private String thumbnail;
+
+    public Theme() {
+    }
 
     public Theme(Long id, String name, String description, String thumbnail) {
-        validate(name, description, thumbnail);
         this.id = id;
         this.name = name;
         this.description = description;
@@ -20,18 +27,6 @@ public class Theme {
 
     public Theme(String name, String description, String thumbnail) {
         this(null, name, description, thumbnail);
-    }
-
-    private void validate(String name, String description, String thumbnail) {
-        if (name == null || name.isBlank()) {
-            throw new InvalidThemeException("테마 이름은 비어있을 수 없습니다.");
-        }
-        if (description == null) {
-            throw new InvalidThemeException("설명은 비어있을 수 없습니다.");
-        }
-        if (thumbnail == null) {
-            throw new InvalidThemeException("썸네일은 비어있을 수 없습니다.");
-        }
     }
 
     public Long getId() {
