+++ src/main/java/roomescape/entity/ThemeRanking.java
@@ -1,10 +1,11 @@
-package roomescape.domain;
+package roomescape.entity;
 
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+// TODO: repo JPA 적용 후, 제거
 public class ThemeRanking {
 
     private static final int RANKING_SIZE = 10;
