+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,9 +3,9 @@ package roomescape.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.Reservation;
-import roomescape.domain.Theme;
-import roomescape.domain.ThemeRanking;
+import roomescape.entity.Reservation;
+import roomescape.entity.Theme;
+import roomescape.entity.ThemeRanking;
 import roomescape.dto.request.CreateThemeRequest;
 import roomescape.exception.InvalidThemeException;
 import roomescape.repository.ReservationRepository;
